backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
supabase>=2.3.4
google-genai>=0.1.0
cohere>=4.47
PyPDF2>=3.0.1
tiktoken>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
